style: rename files for better consistency
